set up input files and start procedure
