+++ PyFluent/configs/static_configs.csv
@@ -1,5 +1,3 @@
-air_density,0.62431
-air_viscosity,1.5781e-05
 residual_continuity,1e-4
 residual_x_velocity,1e-4
 residual_y_velocity,1e-4
@@ -14,7 +12,4 @@ drag_coef_monitor_zone,enclosure-enclosure:1
 drag_coef_monitor_x_vector,0
 drag_coef_monitor_y_vector,-1
 drag_coef_monitor_z_vector,0
-inlet_x_velocity,0
-inlet_y_velocity,-273
-inlet_z_velocity,0
 roughness_constant,5e-6
