+++ PyFluent/configs/variable_configs.csv
@@ -1,0 +1,5 @@
+air_density,1.174739115833875
+air_viscosity,1.7757260613668663e-05
+inlet_x_velocity,75.81486781974664
+inlet_y_velocity,5.301492004553523
+inlet_z_velocity,0.0
