+++ inputs.csv
@@ -1,0 +1,7 @@
+altitude,velocity,angle_of_attack
+100,12,1
+234,23,5
+15,545,6
+235,65,65
+45,67,5
+435,76,4
